refactor : 폴더 구조 수정
app 폴더에 필요한 폴더만 남기고 그 외 폴더들은 src 폴더로 이동
