+++ src/app/components/AuthBtn/AuthBtn.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useAuthStore } from '../../store/authStore';
+import { useAuthStore } from '../../../store/authStore';
 import LogoutBtn from './LogoutBtn';
 import LoginBtn from './LoginBtn';
