+++ src/app/layout.tsx
@@ -10,7 +10,7 @@ export const metadata: Metadata = {
 };
 
 const pretendard = localFont({
-  src: './font/PretendardVariable.woff2',
+  src: '../font/PretendardVariable.woff2',
   display: 'swap',
 });
