+++ src/app/login/page.tsx
@@ -4,10 +4,10 @@ import { Box, Flex, Text } from '@chakra-ui/react';
 import { v4 as uuidv4 } from 'uuid';
 import KakaoLoginBtnSVG from '@/public/assets/KakaoLoginBtn.svg';
 import NaverLoginBtnSVG from '@/public/assets/NaverLoginBtn.svg';
-import { OAUTH } from '@/src/app/config/auth';
+import { OAUTH } from '@/src/config/auth';
 import Footer from '../components/Footer/Footer';
-import { useToast } from '../hooks/useToast';
-import Logo from '../components/Logo/Logo';
+import { useToast } from '../../hooks/useToast';
+import Logo from '../../common/Logo/Logo';
 
 const handleKakaoLogin = () => {
   window.location.href = OAUTH.KAKAO.AUTH_URL;
