+++ src/types/accordionItem.ts
